+++ packages/contract-helpers/src/commons/ipfs.test.ts
@@ -9,7 +9,7 @@ describe('ipfs', () => {
     '0x04d1fd83d352a7caa14408cee133be97b5919c3a5cf79a47ded3c9b658447d79';
   describe('getLink', () => {
     it('Expects the link to be correct', () => {
-      const link = getLink(hash);
+      const link = getLink(hash, 'https://cloudflare-ipfs.com/ipfs');
       expect(link).toEqual(
         'https://cloudflare-ipfs.com/ipfs/0x04d1fd83d352a7caa14408cee133be97b5919c3a5cf79a47ded3c9b658447d79',
       );
@@ -64,7 +64,7 @@ describe('ipfs', () => {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
       mockFetch.mockResolvedValue({ ok: true, json } as Response);
 
-      const metadata = await getProposalMetadata(hash);
+      const metadata = await getProposalMetadata(ipfsHash);
       const metadata2 = await getProposalMetadata(hash);
 
       expect(json.mock.calls.length).toEqual(1);