+++ packages/contract-helpers/src/governance-contract/governance.test.ts
@@ -1,5 +1,4 @@
 import { BigNumber, providers, utils } from 'ethers';
-import * as ipfs from '../commons/ipfs';
 import { eEthereumTxType, GasType, transactionType } from '../commons/types';
 import { IAaveGovernanceV2 } from './typechain/IAaveGovernanceV2';
 import { IAaveGovernanceV2__factory } from './typechain/IAaveGovernanceV2__factory';
@@ -180,6 +179,52 @@ describe('GovernanceService', () => {
       );
     });
   });
+  describe('getProposal', () => {
+    afterEach(() => {
+      jest.clearAllMocks();
+    });
+    it('Expects a proposal parsed if all params passed correctly', async () => {
+      const instance = new AaveGovernanceService(provider, {
+        GOVERNANCE_ADDRESS,
+        GOVERNANCE_HELPER_ADDRESS,
+      });
+      const spy = jest
+        .spyOn(IGovernanceV2Helper__factory, 'connect')
+        .mockReturnValue({
+          getProposal: async () => Promise.resolve(proposalMock),
+        } as unknown as IGovernanceV2Helper);
+
+      const proposal = await instance.getProposal({ proposalId: 1 });
+
+      expect(spy).toHaveBeenCalled();
+      expect(proposal).toEqual({
+        id: Number(proposalMock.id.toString()),
+        creator: proposalMock.creator,
+        executor: proposalMock.executor,
+        targets: proposalMock.targets,
+        values: proposalMock.values,
+        signatures: proposalMock.signatures,
+        calldatas: proposalMock.calldatas,
+        withDelegatecalls: proposalMock.withDelegatecalls,
+        startBlock: Number(proposalMock.startBlock.toString()),
+        endBlock: Number(proposalMock.endBlock.toString()),
+        executionTime: proposalMock.executionTime.toString(),
+        forVotes: proposalMock.forVotes.toString(),
+        againstVotes: proposalMock.againstVotes.toString(),
+        executed: proposalMock.executed,
+        canceled: proposalMock.canceled,
+        strategy: proposalMock.strategy,
+        ipfsHash: proposalMock.ipfsHash,
+        state: Object.values(ProposalState)[proposalMock.proposalState],
+        minimumQuorum: proposalMock.minimumQuorum.toString(),
+        minimumDiff: proposalMock.minimumDiff.toString(),
+        executionTimeWithGracePeriod:
+          proposalMock.executionTimeWithGracePeriod.toString(),
+        proposalCreated: Number(proposalMock.proposalCreated.toString()),
+        totalVotingSupply: proposalMock.totalVotingSupply.toString(),
+      });
+    });
+  });
   describe('getProposals', () => {
     afterEach(() => {
       jest.clearAllMocks();
@@ -191,18 +236,6 @@ describe('GovernanceService', () => {
         GOVERNANCE_ADDRESS,
         GOVERNANCE_HELPER_ADDRESS,
       });
-      const ipfsSpy = jest.spyOn(ipfs, 'getProposalMetadata').mockReturnValue(
-        Promise.resolve({
-          title: 'mockTitle',
-          description: 'mockDescription',
-          shortDescription: 'mockShortDescription',
-          ipfsHash:
-            '0x04d1fd83d352a7caa14408cee133be97b5919c3a5cf79a47ded3c9b658447d79',
-          aip: 0,
-          author: 'lachs',
-          discussions: 'link',
-        }),
-      );
       const spy = jest
         .spyOn(IGovernanceV2Helper__factory, 'connect')
         .mockReturnValue({
@@ -212,7 +245,6 @@ describe('GovernanceService', () => {
       const proposals = await instance.getProposals({ skip, limit });
 
       expect(spy).toHaveBeenCalled();
-      expect(ipfsSpy).toHaveBeenCalled();
       expect(proposals[0]).toEqual({
         id: Number(proposalMock.id.toString()),
         creator: proposalMock.creator,
@@ -236,19 +268,14 @@ describe('GovernanceService', () => {
         minimumDiff: proposalMock.minimumDiff.toString(),
         executionTimeWithGracePeriod:
           proposalMock.executionTimeWithGracePeriod.toString(),
-        title: 'mockTitle',
-        description: 'mockDescription',
-        shortDescription: 'mockShortDescription',
         proposalCreated: Number(proposalMock.proposalCreated.toString()),
         totalVotingSupply: proposalMock.totalVotingSupply.toString(),
-        aip: 0,
-        author: 'lachs',
-        discussions: 'link',
       });
     });
     it('Expects to fail if gov address not eth address', async () => {
       const instance = new AaveGovernanceService(provider, {
         GOVERNANCE_ADDRESS: 'asdf',
+        ipfsGateway: 'https://cloudflare-ipfs.com/ipfs',
       });
       const getProposals = instance.getProposals({ skip, limit });
       expect(getProposals).toEqual([]);